disables if the bots perms get revoked

+++ cogs/anti_spam.py
@@ -1,6 +1,6 @@
 from discord.ext import commands
 from os.path import isfile
-from utils import colors
+from utils import colors, utils
 from time import time
 import discord
 import asyncio
@@ -11,7 +11,7 @@ class Anti_Spam(commands.Cog):
 		self.bot = bot
 		self.toggle = {}
 		self.roles = {}
-		self.working = {}
+		self.status = {}
 		self.cd = {}
 		if isfile("./data/userdata/anti_spam.json"):
 			with open("./data/userdata/anti_spam.json", "r") as f:
@@ -64,9 +64,11 @@ class Anti_Spam(commands.Cog):
 	@commands.Cog.listener()
 	async def on_message(self, m: discord.Message):
 		if isinstance(m.guild, discord.Guild):
+			if "spam" in m.channel.name:
+				return
 			guild_id = str(m.guild.id)
 			if guild_id in self.toggle:
-				if m.author.bot:
+				if m.author.id is self.bot.user.id:
 					return
 				user_id = str(m.author.id)
 				now = int(time() / 5)
@@ -79,46 +81,55 @@ class Anti_Spam(commands.Cog):
 				else:
 					self.cd[guild_id][user_id] = [now, 0]
 				if self.cd[guild_id][user_id][1] > 2:
+					bot = m.guild.get_member(self.bot.user.id)
+					perms = [perm for perm, value in bot.guild_permissions]
+					if "manage_roles" not in perms:
+						del self.toggle[guild_id]
+						return
 					if m.author.top_role.position >= m.guild.get_member(self.bot.user.id).top_role.position:
 						return await m.delete()
-					if guild_id not in self.working:
-						self.working[guild_id] = {}
-					if user_id in self.working[guild_id]:
+					with open("./data/userdata/mod.json", "r") as f:
+						dat = json.load(f)  # type: dict
+						if "timers" in dat:
+							if user_id in dat['timers']:
+								return
+					if guild_id not in self.status:
+						self.status[guild_id] = {}
+					if user_id in self.status[guild_id]:
 						return
-					self.working[guild_id][user_id] = "working"
-					role = discord.utils.get(m.guild.roles, name="Muted")
-					if not role:
-						role = discord.utils.get(m.guild.roles, name="muted")
-					if not role:
-						role = await m.guild.create_role(name="Muted", color=discord.Color(colors.black()), hoist=True)
+					self.status[guild_id][user_id] = "working"
+					mute_role = discord.utils.get(m.guild.roles, name="Muted")
+					if not mute_role:
+						mute_role = discord.utils.get(m.guild.roles, name="muted")
+					if not mute_role:
+						mute_role = await m.guild.create_role(name="Muted", color=discord.Color(colors.black()), hoist=True)
 						for channel in m.guild.text_channels:
-							await channel.set_permissions(role, send_messages=False)
+							await channel.set_permissions(mute_role, send_messages=False)
 						for channel in m.guild.voice_channels:
-							await channel.set_permissions(role, speak=False)
+							await channel.set_permissions(mute_role, speak=False)
 					roles = []
-					for r in m.author.roles:
+					for role in m.author.roles:
 						try:
-							await m.author.remove_roles(r)
-							roles.append(r.id)
-							await asyncio.sleep(0.5)
+							await m.author.remove_roles(role)
+							roles.append(role.id)
+							await asyncio.sleep(1)
 						except:
 							pass
-					await m.author.add_roles(role)
-					try:
+					await m.author.add_roles(mute_role)
+					if utils.Bot().can_dm(user=m.author):
 						await m.author.send(f"You've been muted for spam in **{m.guild.name}** for 2mins and 30secs")
-					except:
-						pass
 					await asyncio.sleep(150)
-					rs = []
-					for r in m.author.roles:
-						rs.append(r.name.lower())
-					if "muted" not in rs:
-						return
-					await m.author.remove_roles(role)
-					for r in roles:
-						await m.author.add_roles(m.guild.get_role(r))
-						await asyncio.sleep(0.5)
-					del self.working[guild_id][user_id]
+					with open("./data/userdata/mod.json", "r") as f:
+						dat = json.load(f)  # type: dict
+						if "timers" in dat:
+							if user_id in dat['timers']:
+								return
+					if mute_role in m.author.roles:
+						await m.author.remove_roles(mute_role)
+						for role in roles:
+							await m.author.add_roles(m.guild.get_role(role))
+							await asyncio.sleep(1)
+						del self.status[guild_id][user_id]
 
 	@commands.Cog.listener()
 	async def on_guild_remove(self, guild):
